# src/controllers/v3/openapi.yaml
components:
...
    schemas:
        OrganizationRefID:
            type: string
        OrganizationUnitRefID:
            type: string
        Dataset:
            properties:
                extForeignKey:
                    type: string
                name:
                    type: string
                license_title:
                    type: string
                security_class:
                    type: string
                view_audience:
                    type: string
                download_audience:
                    type: string
                record_publish_date:
                    type: string
                notes:
                    type: string
                title:
                    type: string
                isInCatalog:
                    type: string
                isDraft:
                    type: string
                contacts:
                    type: string
                extSource:
                    type: string
                extRecordHash:
                    type: string
                tags:
                    items:
                        type: string
                    type: array
                resources: {}
                organization:
                    $ref: '#/components/schemas/OrganizationRefID'
                organizationUnit:
                    $ref: '#/components/schemas/OrganizationUnitRefID'
            type: object
            additionalProperties: false
        BatchResult:
            properties:
                status:
                    type: number
                    format: double
                result:
                    type: string
                reason:
                    type: string
                id:
                    type: string
                ownedBy:
                    type: string
                childResults:
                    items:
                        $ref: '#/components/schemas/BatchResult'
                    type: array
            required:
                - status
                - result
            type: object
            additionalProperties: false
        DraftDataset:
            properties:
                name:
                    type: string
                license_title:
                    type: string
                security_class:
                    type: string
                    enum:
                        - HIGH-CABINET
                        - HIGH-CONFIDENTIAL
                        - HIGH-SENSITIVITY
                        - MEDIUM-SENSITIVITY
                        - MEDIUM-PERSONAL
                        - LOW-SENSITIVITY
                        - LOW-PUBLIC
                        - PUBLIC
                        - 'PROTECTED A'
                        - 'PROTECTED B'
                        - 'PROTECTED C'
                view_audience:
                    type: string
                    enum:
                        - Public
                        - Government
                        - 'Named users'
                        - 'Government and Business BCeID'
                download_audience:
                    type: string
                    enum:
                        - Public
                        - Government
                        - 'Named users'
                        - 'Government and Business BCeID'
                record_publish_date:
                    type: string
                notes:
                    type: string
                title:
                    type: string
                isInCatalog:
                    type: boolean
                isDraft:
                    type: boolean
                contacts:
                    type: string
                resources:
                    type: string
                tags:
                    items:
                        type: string
                    type: array
                organization:
                    $ref: '#/components/schemas/OrganizationRefID'
                organizationUnit:
                    $ref: '#/components/schemas/OrganizationUnitRefID'
            type: object
            additionalProperties: false
            example:
                name: my_sample_dataset
                license_title: 'Open Government Licence - British Columbia'
                security_class: PUBLIC
                view_audience: Public
                download_audience: Public
                record_publish_date: '2017-09-05'
                notes: 'Some notes'
                title: 'A title about my dataset'
                tags:
                    - tag1
                    - tag2
                organization: ministry-of-citizens-services
                organizationUnit: databc
        Gateway:
            properties:
                gatewayId:
                    type: string
                displayName:
                    type: string
            type: object
            additionalProperties: false
...
    securitySchemes:
        jwt:
            type: oauth2
            description: 'Authz Client Credential'
            flows:
                clientCredentials:
                    tokenUrl: 'https://token_endpoint'
                    scopes: {}
        portal:
            type: http
            description: 'Authz Portal Login'
            scheme: bearer
            bearerFormat: JWT
        openid:
            type: openIdConnect
            description: 'OIDC Login'
            openIdConnectUrl: 'https://well_known_endpoint'
info:
    title: 'APS Directory API'
    version: 3.0.0
    description: 'API Services Portal by BC Gov API Programme Services'
    license:
        name: MIT
    contact:
        name: 'BC Gov APS'
openapi: 3.0.0
paths:
    '/organizations/{org}/datasets':
        get:
            operationId: organization-datasets
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Dataset'
                                type: array
            description: "Get metadata about Datasets that are available by API for this organization\n> `Required Scope:` Dataset.Manage"
            summary: 'Get Organization Datasets'
            tags:
                - 'API Directory'
            security:
                -
                    jwt:
                        - Dataset.Manage
            parameters:
                -
                    in: path
                    name: org
                    required: true
                    schema:
                        type: string
        put:
            operationId: put-organization-dataset
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/BatchResult'
            description: "Manage metadata about Datasets that are available by API for this organization\n> `Required Scope:` Dataset.Manage"
            summary: 'Manage Organization Datasets'
            tags:
                - 'API Directory'
            security:
                -
                    jwt:
                        - Dataset.Manage
            parameters:
                -
                    in: path
                    name: org
                    required: true
                    schema:
                        type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/DraftDataset'
    '/organizations/{org}/datasets/{name}':
        delete:
            operationId: delete-dataset
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/BatchResult'
            description: "Delete a Dataset\n> `Required Scope:` Dataset.Manage"
            summary: 'Delete a dataset'
            tags:
                - 'API Directory'
            security:
                -
                    jwt:
                        - Dataset.Manage
            parameters:
                -
                    in: path
                    name: org
                    required: true
                    schema:
                        type: string
                -
                    in: path
                    name: name
                    required: true
                    schema:
                        type: string
        get:
            operationId: get-organization-dataset
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Dataset'
            description: "Get metadata about a Dataset that are available by API for this organization\n> `Required Scope:` Dataset.Manage"
            summary: 'Get Organization Dataset'
            tags:
                - 'API Directory'
            security:
                -
                    jwt:
                        - Dataset.Manage
            parameters:
                -
                    in: path
                    name: org
                    required: true
                    schema:
                        type: string
                -
                    in: path
                    name: name
                    required: true
                    schema:
                        type: string
...
    '/directory/{id}':
        get:
            operationId: directory-item
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema: {}
            tags:
                - 'API Directory'
            security: []
            parameters:
                -
                    in: path
                    name: id
                    required: true
                    schema:
                        type: string
...
    '/gateways/{gatewayId}':
        get:
            operationId: namespace-profile
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Gateway'
            description: "Get details about the gateway, such as permissions for what the gateway is setup with.\n> `Required Scope:` Namespace.Manage"
            summary: 'Gateway Summary'
            tags:
                - Gateways
            security:
                -
                    jwt:
                        - Namespace.Manage
            parameters:
                -
                    in: path
                    name: gatewayId
                    required: true
                    schema:
                        type: string
        delete:
            operationId: delete-namespace
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Gateway'
            description: "Delete the gateway\n> `Required Scope:` Namespace.Manage"
            summary: 'Delete Gateway'
            tags:
                - Gateways
            security:
                -
                    jwt:
                        - Namespace.Manage
            parameters:
                -
                    in: path
                    name: gatewayId
                    required: true
                    schema:
                        type: string
                -
                    in: query
                    name: force
                    required: false
                    schema:
                        default: false
                        type: boolean
...
    '/gateways/{gatewayId}/directory/{id}':
        get:
            operationId: get-ns-directory-dataset
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema: {}
            description: "Used primarily for \"Preview Mode\"\nGet a particular Dataset"
            tags:
                - 'API Directory'
            security:
                -
                    jwt:
                        - Namespace.Manage
            parameters:
                -
                    in: path
                    name: gatewayId
                    required: true
                    schema:
                        type: string
                -
                    in: path
                    name: id
                    required: true
                    schema:
                        type: string
    '/gateways/{gatewayId}/directory':
        get:
            operationId: get-ns-directory
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema: {}
            description: "Used primarily for \"Preview Mode\"\nList the datasets belonging to a particular namespace"
            tags:
                - 'API Directory'
            security:
                -
                    jwt:
                        - Namespace.Manage
            parameters:
                -
                    in: path
                    name: gatewayId
                    required: true
                    schema:
                        type: string
...
    '/organizations/{org}/{orgUnit}/gateways/{gatewayId}':
        put:
            operationId: assign-namespace-to-organization
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    result: {type: string}
                                required:
                                    - result
                                type: object
            description: '> `Required Scope:` Namespace.Assign'
            tags:
                - Organizations
            security:
                -
                    jwt:
                        - Namespace.Assign
            parameters:
                -
                    in: path
                    name: org
                    required: true
                    schema:
                        type: string
                -
                    in: path
                    name: orgUnit
                    required: true
                    schema:
                        type: string
                -
                    in: path
                    name: gatewayId
                    required: true
                    schema:
                        type: string
                -
                    in: query
                    name: enable
                    required: false
                    schema:
                        default: true
                        type: boolean
        delete:
            operationId: unassign-namespace-from-organization
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    result: {type: string}
                                required:
                                    - result
                                type: object
            description: '> `Required Scope:` Namespace.Assign'
            tags:
                - Organizations
            security:
                -
                    jwt:
                        - Namespace.Assign
            parameters:
                -
                    in: path
                    name: org
                    required: true
                    schema:
                        type: string
                -
                    in: path
                    name: orgUnit
                    required: true
                    schema:
                        type: string
                -
                    in: path
                    name: gatewayId
                    required: true
                    schema:
                        type: string
...
servers:
    -
        url: /ds/api/v3
...